% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.save.R
\docType{data}
\name{data}
\alias{data}
\title{result.save}
\format{
An object of class \code{numeric} of length 6.
}
\usage{
data
}
\arguments{
\item{data}{the result data to be saved}

\item{id}{unique ID of this result. If not specified, a unique Id is randomly generated and returned.}

\item{resultvar}{result variable to save to}

\item{label}{designation of this result in the research article (e.g., 'Table 2')}

\item{type}{type of result, i.e., 'Regression', 'Pairwise t-Test', etc.}

\item{description}{longer description of the result, e.g., 'Test for a treatment effect in the pooled data.'}
}
\description{
Saves a result (a number, text, vector, list, table, plot, etc.) with proper documentation.
}
\examples{
Instead of writing, for example:
A[A$a==B$a & A$b==B$b & A$c==B$c & A$d==B$d,]
one can write
A[SPSubset(A,B,c("a","b","c","d")),]
}
\keyword{documentation}
\keyword{results,}
