% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.save.R
\name{result.save}
\alias{result.save}
\title{result.save}
\usage{
result.save(
  data = NULL,
  id = NA,
  resultvar = "Result",
  label = NA,
  type = NA,
  description = NA,
  id.n = 4,
  use_lowercase = T,
  use_uppercase = F,
  use_numeric = T
)
}
\arguments{
\item{data}{the result data to be saved}

\item{id}{unique ID of this result. If not specified, a unique Id is randomly generated and returned.}

\item{resultvar}{result variable to save to}

\item{label}{designation of this result in the research article (e.g., 'Table 2')}

\item{type}{type of result, i.e., 'Regression', 'Pairwise t-Test', etc.}

\item{description}{longer description of the result, e.g., 'Test for a treatment effect in the pooled data.'}

\item{id.n}{the number of random ID characters.}

\item{use_lowercase}{a logical value indicating whether lower case letters 
should be used for the ID}

\item{use_uppercase}{a logical value indicating whether upper case letters 
should be used for the ID}

\item{use_numeric}{a logical value indicating whether numeric digits' should
be used for the ID}
}
\description{
Saves a result (a number, text, vector, list, table, plot, etc.) with proper documentation.
}
\examples{
result.save(table(x),
            label = "Table 1",
            type = "Summary table",
            description = "Table listing the experimental conditions.")
}
\keyword{documentation}
\keyword{results,}
