% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.new.id.R
\name{result.new.id}
\alias{result.new.id}
\title{result.new.id}
\usage{
result.new.id(
  n = 4,
  resultvar,
  use_lowercase = T,
  use_uppercase = F,
  use_numeric = T
)
}
\arguments{
\item{n}{number of random ID characters}

\item{resultvar}{result variable to check for ID uniqueness}
}
\description{
Draws a new, unique, random result ID.
}
\examples{
result.new.id()
result.new.id(n = 4, resultvar = "Result")
Creates a new, random, 4-character ID that is unique among all IDs in the 'Result' results object
result.new.id(6, resultvar = "main", use_uppercase = T, use_numeric = F)
Creates a new, random, 6-character ID that is unique among all IDs in the 'main' results object
}
\keyword{ID}
\keyword{results,}
\keyword{unique}
